version 1.7.1
git-svn-id: https://plugins.svn.wordpress.org/better-anchor-links/trunk@1017543 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ auto-anchor-list.php
@@ -4,7 +4,7 @@ Plugin Name: Better Anchor Links
 Plugin URI: http://ludek.org/bal/index.html
 Description: Automatically creates and displays anchor links.
 Author: Luděk Melichar
-Version: 1.7
+Version: 1.7.1
 Author URI: http://ludek.org
 */
 
@@ -29,7 +29,7 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 if (!class_exists('mwm_aalLoader')) {
 	class mwm_aalLoader{
 		
-		var $version     = '1.7';
+		var $version     = '1.7.1';
 		var $options     = '';
 		var $links = array();
 	
@@ -86,6 +86,12 @@ if (!class_exists('mwm_aalLoader')) {
 				{
 					$lm_bal_options=$options ;
 					$lm_bal_options['backlink_char'] = "*";
+					update_option('lm_bal_options', $lm_bal_options); 
+				}
+				if (!array_key_exists('is_backlinkfront', $options))
+				{
+					$lm_bal_options=$options ;
+					$lm_bal_options['is_backlinkfront'] = false;
 					update_option('lm_bal_options', $lm_bal_options); 
 				}
 			}
